move menu code to library
